Delete deprecated class TableExtension
The logic from this class was moved into the individual table classes.
Also comment out some tests that depended on this deleted table extension
class.

Unfortunately the all the tests haven't been maintained after different
refactorings, so it's still the case that they don't all work. But I think
that is a task that we should postpone until @aszepieniec feels that
this STARK engine has been completed for the tutorial.