Fix error in symbolic degree bounds evaluation for mpols
When the coefficient is zero for a term in a multivariate polynomium
this term must not count towards the degree of the evaluated univariate
polynomium degree.

Two tests have been added, and the entire STARK Brainfuck engine for
proof + verify have been verified to still work with four programs:
`++++`, Hello World, `,.,.,{.x17}`, and a program that prints 17 '!'.

All four programs worked before and after this change, so this error
didn't seem to affect the completeness of the proofs.